// packages/expo-modules-autolinking/src/types.ts
export interface HarmonyHarConfig {
  packageName: string;
  packagePath: string;
  primary?: boolean;
  version?: string;
  transform?: 'screens-content-wrapper-v1' | 'worklets-private-symbols-v2';
}